rename {Interpolation => interpolation}.hpp
